CaimCore API

Core

CaimCore.version — Function

version()

Get the current version of CaimCore.

julia> CaimCore.version()
v"0.1.0"

Sessions

Sessions provide the core organizational structure of CaimCore, connecting the various phases of analysis, and providing persistent storage.

CaimCore.Session — Type

Session(path[, storage])

A structure for storing session information for a CaimCore analysis.

CaimCore.SessionStorage — Type

SessionStorage([version])

Persistent storage component of a CaimCore Session. This data will be saved and read from disk via calls to save and load.

julia> SessionStorage()
SessionStorage(v"0.1.0")

CaimCore.save — Method

save(filename, session)

Save the persistent component of a Session to a disc.

CaimCore.load — Method

load(::Type{Session}, filename)

Load a Session from disc.

Datasets

CaimCore.Dataset — Type

Dataset

A supertype for all loadable datasets

CaimCore.VideoGraphicDataset — Type

VideoGraphicDataset <: Dataset

A supertype for all videographic datasets

Image Stacks

CaimCore.ImageStack — Type

ImageStack <: VideoGraphicDataset
ImageStack(frames::Array{C,3}) where {C <: Colorant}
ImageStack(frames::AbstractVector{Matrix{C}} where {C <: Colorant}

An image stack consists of a collection of 2-D images, running in time.

CaimCore.load — Method

load(::Type{ImageStack}, dir[, ext]; ignoredots=true)

Load all image files from a directory into an ImageStack. If provided, only files with extention ext will be loaded. Files whose name begins with a . will be ignored if ignoredots is true.

The files are sorted lexicographically before being loaded. The files must all have the same size, file extention and color space, e.g. RGB{N0f8}, Gray{N0f8}, etc..

julia> readdir()
6-element Array{String,1}:
 ".DS_Store"
 "000001.tiff"
 "000002.tiff"
 "000003.tiff"
 "000004.tiff"
 "000005.tiff"

julia> load(ImageStack, ".")
ImageStack{ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}}
  size: 20×20×5

julia> load(ImageStack, "."; ignoredots=false)
ERROR: files must all have the same extension; got ["", ".tiff"]
[...]

You can filter files by extension.

julia> load(ImageStack, ".")
ERROR: files must all have the same extension; got [".png", ".tiff"]
[...]

julia> load(ImageStack, ".", ".tiff")
ImageStack{ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}}
  size: 20×20×2

CaimCore.load — Method

load(::Type{ImageStack}, files)

Load files into an image stack in the order they are provided.

The files must all have the same size, file extention and color space, e.g. RGB{N0f8}, Gray{N0f8}, etc..

julia> load(ImageStack, ["000001.tiff", "000002.tiff"])
ImageStack{ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}}
  size: 20×20×2