CaimCore API

Core

CaimCore.version — Function

version()

Get the current version of CaimCore.

julia> CaimCore.version()
v"0.1.0"

Sessions

Sessions provide the core organizational structure of CaimCore, connecting the various phases of analysis, and providing persistent storage.

CaimCore.Session — Type

Session(path[, storage])

A structure for storing session information for a CaimCore analysis.

CaimCore.SessionStorage — Type

SessionStorage([version])

Persistent storage component of a CaimCore Session. This data will be saved and read from disk via calls to save and load.

julia> SessionStorage()
SessionStorage(v"0.1.0")

CaimCore.save — Function

save(filename, session)

Save the persistent component of a Session to a disc.

CaimCore.load — Function

load(::Type{Session}, filename)

Load a Session from disc.